server/requirements.txt
appdirs==1.4.4
asn1crypto==1.4.0
audioread==2.1.9
autopep8==1.5.7
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cycler==0.10.0
decorator==5.0.9
Flask==1.1.2
Flask-Cors==3.0.10
google-api-core==1.30.0
google-auth==1.32.0
google-cloud-speech==2.4.1
googleapis-common-protos==1.53.0
greenlet==1.0.0
grpcio==1.38.1
gunicorn==20.1.0
idna==2.10
install==1.3.4
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
kiwisolver==1.3.1
libcst==0.3.19
librosa==0.8.1
llvmlite==0.36.0
MarkupSafe==1.1.1
matplotlib==3.4.1
mecab==0.996.2
mecab-python==1.0.0
mecab-python3==1.0.4
mido==1.2.10
module==0.0.4
mojimoji==0.0.11
mypy-extensions==0.4.3
neologdn==0.4
numba==0.53.1
numpy==1.20.2
packaging==20.9
pandas==1.2.3
pg8000==1.19.5
Pillow==8.2.0
pooch==1.4.0
pretty-midi==0.2.9
proto-plus==1.18.1
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pydub==0.25.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
resampy==0.2.2
rsa==4.7.2
scikit-learn==0.24.1
scipy==1.6.2
scramp==1.4.0
six==1.15.0
SoundFile==0.10.3.post1
SpeechRecognition==3.8.1
spotipy==2.18.0
SQLAlchemy==1.4.13
threadpoolctl==2.1.0
toml==0.10.2
typing-extensions==3.10.0.0
typing-inspect==0.7.1
urllib3==1.26.4
Werkzeug==1.0.1
